requirements.txt
streamlit
python-dotenv
youtube-transcript-api==0.6.2
pytube

sqlalchemy
psycopg2-binary

redis

pinecone

langchain
langchain-openai
tiktoken

ragas
pandas
numpy

langsmith

yt-dlp
webvtt-py

